% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_n_lag_columns}
\alias{add_n_lag_columns}
\title{Add lagged versions of column \code{var} to \code{dat}}
\usage{
add_n_lag_columns(dat, var, n_lags)
}
\arguments{
\item{dat}{Data frame with at least one column, \code{var}.}

\item{var}{Column in \code{dat} that will be lagged.}

\item{n_lags}{Vector of lagged versions of \code{var} to include. For
example, if \code{n_lags} = 1, a lag-1 version named \code{var_lag1} will
be added to \code{dat}. If \code{n_lags} = 2:3, the lag-2 and lag-3
versions will be added and named \code{var_lag2} and \code{var_lag3},
respectively.}
}
\value{
Returns \code{dat} with additional columns that are lagged versions
of column \code{var}.
}
\description{
This is an internal function used in \code{qc_test_flat_line()}.
}
