% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assert_county}
\alias{assert_county}
\title{Ensure there is one and only one county specified for the qc_test_* functions}
\usage{
assert_county(dat, county_arg, foo)
}
\arguments{
\item{dat}{Data frame sent to the qc_test_* function.}

\item{county_arg}{Character string (or \code{NULL} value) in the county
argument of the qc_test_*() function.}

\item{foo}{Character string indicating which qc_test_* function is being
checked.}
}
\value{
Returns a character string indicating the county for which to apply
thresholds.
}
\description{
Ensure there is one and only one county specified for the qc_test_* functions
}
