% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{ggplot_all_tests}
\alias{ggplot_all_tests}
\title{Title}
\usage{
ggplot_all_tests(dat, qc_test, var)
}
\arguments{
\item{dat}{placeholder}

\item{qc_test}{qc test to plot}

\item{var}{variable to plot}
}
\value{
ggplot object
}
\description{
Title
}
