% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tables.R
\docType{data}
\name{months_seasons}
\alias{months_seasons}
\title{Seasons tables}
\format{
Months assigned to seasons.
}
\source{
Add references here
}
\usage{
months_seasons
}
\description{
Seasons tables
}
\keyword{datasets}
