% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot.R
\name{pivot_flags_longer}
\alias{pivot_flags_longer}
\title{Title}
\usage{
pivot_flags_longer(dat_wide, qc_test)
}
\arguments{
\item{dat_wide}{placeholder...}

\item{qc_test}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
