% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot_wider.R
\name{pivot_flags_wider}
\alias{pivot_flags_wider}
\title{Complete pivot_wider of flagged sensor string data}
\usage{
pivot_flags_wider(dat_long, var, var_position = NULL)
}
\arguments{
\item{dat_long}{Data frame of flagged sensor string data with the variables
in a long format and flags in wide format.}

\item{var}{Variable to pivot.}

\item{var_position}{Character string. Name of the column in \code{dat_long}
that the the variable column will be located after. By default, each
variable column will be located after the column
\code{sensor_depth_at_low_tide_m}.}
}
\value{
Returns \code{dat_long} with the variable and qc_test flag columns
pivoted to a wide format.
}
\description{
Complete pivot_wider of flagged sensor string data
}
