% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot.R
\name{pivot_test_longer}
\alias{pivot_test_longer}
\title{Title}
\usage{
pivot_test_longer(dat, qc_test)
}
\arguments{
\item{dat}{placeholder}

\item{qc_test}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
