% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_assemble_county_datasets.R
\name{qc_assemble_county_data}
\alias{qc_assemble_county_data}
\title{Assemble county data}
\usage{
qc_assemble_county_data(prov = "ns", path = NULL, folder)
}
\arguments{
\item{prov}{Character string indicating which province the data to be
assembled is from. Options are "ns" (the default) and "nb". This dictates
the default file path.}

\item{path}{File path the the folder with the rds files to be assembled.}

\item{folder}{Name of the folder where the rds files are saved.}
}
\value{
Returns a data.frame with data from all deployments in folder.
}
\description{
Will add columns that do not already exist.
}
