% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_flag_labels.R
\name{qc_assign_flag_labels}
\alias{qc_assign_flag_labels}
\title{Assign text labels to flag values}
\usage{
qc_assign_flag_labels(dat)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long or wide format.
Must include at least one column name with the string "flag".}
}
\value{
Returns \code{dat} with entries in the flag columns converted from
numeric values to text labels (an ordered factor with levels "Pass", "Not
Evaluated", "Suspect", "Fail", "Missing Data").
}
\description{
All columns with the string "flag" in the name will be converted from numeric
flags to text flags (e.g., 1 = Pass, 2 = Not Evaluated, 3 = Suspect, 4 =
Fail).
}
\examples{
path <- system.file("testdata", package = "qaqcmar")
dat <- readRDS(paste0(path, "/test_data.RDS"))

dat_qc <- dat \%>\%
 qc_test_grossrange() \%>\%
 qc_assign_flag_labels()
}
