% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_assign_max_flag.R
\name{qc_assign_max_flag2}
\alias{qc_assign_max_flag2}
\title{Assign each observation the maximum flag from applied QC tests.}
\usage{
qc_assign_max_flag2(dat, qc_tests = NULL)
}
\arguments{
\item{dat}{Data frame in long or wide format with flag columns from multiple
quality control tests.}

\item{qc_tests}{Quality control tests included in \code{dat}.}
}
\value{
Returns \code{dat} in a wide format, with a single flag column for
each variable column.
}
\description{
** change the spike test values at beginning/end to NA instead of 2
}
