% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calculate_thresholds.R
\name{qc_calculate_depth_crosscheck_thresholds}
\alias{qc_calculate_depth_crosscheck_thresholds}
\title{Calculate depth cross check thresholds}
\usage{
qc_calculate_depth_crosscheck_thresholds(dat, prob = 0.95)
}
\arguments{
\item{dat}{Data frame. Must include columns
\code{sensor_depth_at_low_tide_m}, \code{sensor_depth_measured_m},
\code{county}, \code{station}, \code{deployment_range}, and
\code{sensor_serial_number}.}

\item{prob}{Quantile to use for the threshold. Sent to \code{quantile()}.}
}
\value{
Tibble with columns \code{qc_test}, \code{variable},
\code{threshold}, and \code{threshold_value}.
}
\description{
Calculate depth cross check thresholds
}
