% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calculate_thresholds.R
\name{qc_calculate_spike_thresholds}
\alias{qc_calculate_spike_thresholds}
\title{Calculate spike test thresholds}
\usage{
qc_calculate_spike_thresholds(
  dat,
  var,
  ...,
  stat = NULL,
  prob = 0.95,
  n_sd = 3,
  fail_factor = 3
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format. Must include
columns \code{county}, \code{station}, \code{deployment_range}, var, and
\code{sensor_serial_number}.}

\item{var}{Unquoted character string indicating which column in \code{dat} to
calculate the thresholds for, e.g., \code{var = temperature_degree_c}.}

\item{...}{Optional grouping variable(s).}

\item{stat}{Statistic to calculate for the threshold. Options are: 1.
"quartile", which uses the quantile function to calculate the value of the
\code{prob} quartile. Used for variables with a right-tailed distribution
for rolling standard deviation.
\enumerate{
\item "mean_sd", which calculates the threshold as the mean + \code{n_sd}
standard deviations. Used for variables with normal distributions for
rolling standard deviation.
}}

\item{prob}{Quantile to use for the threshold. Sent to \code{quantile()}.}

\item{n_sd}{Number of standard deviations to add / subtract from mean to
calculate the user thresholds. Default is \code{n_sd = 3}.}

\item{fail_factor}{Numeric. Used to calculate \code{spike_high} based on
\code{spike_low}: \code{spike_high} = \code{max_max} * \code{spike_low}.}
}
\value{
Tibble with columns \code{qc_test}, \code{variable},
\code{threshold}, and \code{threshold_value}.
}
\description{
Returns 2 levels of spike threshold: \code{spike_low} is calculated from
historical data and the selected statistics. Values greater than
\code{spike_low} will be flagged as Suspect / Of Interest.
}
\details{
\code{spike_high} is a user-specified multiple of \code{spike_low}. Values
greater than \code{spike_high} will be flagged Fail.
}
