% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calculate_thresholds.R
\name{qc_calculate_user_thresholds}
\alias{qc_calculate_user_thresholds}
\title{Calculate gross range test user thresholds}
\usage{
qc_calculate_user_thresholds(dat, var, n_sd = 3, keep_stats = FALSE)
}
\arguments{
\item{dat}{Data frame with at least one column.}

\item{var}{Unquoted character string indicating which column in \code{dat} to
calculate the thresholds for, e.g., \code{var = temperature_degree_c}.}

\item{n_sd}{Number of standard deviations to add / subtract from mean to
calculate the user thresholds. Default is \code{n_sd = 3}.}

\item{keep_stats}{Logical argument indicating whether to keep the calculated
mean and standard deviation columns.}
}
\value{
Returns a tibble with county, variable, user_min, user_max (and
optionally mean_var and sd_var).
}
\description{
The gross range user thresholds are calculated from historical data as the
mean +/- 3 standard deviations:
}
\details{
\deqn{user_{min} = avg_{var} - 3 * stdev_{var}} \deqn{user_{max} = avg_{var} +
3 * stdev_{var}}

where \eqn{avg_{var}} is average of the variable of interest, and
\eqn{stdev_{var}} is the standard deviation of the variable of interest.

This function does not automatically group by any variables. To calculate the
user thresholds by group (e.g., county) use
\code{dplyr::group_by(group_variable)}, and send the results to
\code{qc_calculate_user_thresholds()}.
}
