% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_flag_range.R
\name{qc_flag_range}
\alias{qc_flag_range}
\title{Add flag column for the gap test}
\usage{
qc_flag_range(dat, variable, threshold)
}
\arguments{
\item{dat}{Data.frame with at least one column \code{timestamp}.}

\item{variable}{Variable to apply the test to.}

\item{threshold}{Gap that fails test.}
}
\value{
placeholder for now
}
\description{
Add flag column for the gap test
}
