% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_flag_labels.R
\name{qc_make_flag_labels}
\alias{qc_make_flag_labels}
\title{Make text labels from flag values}
\usage{
qc_make_flag_labels(dat)
}
\arguments{
\item{dat}{Dataframe with at least one column with "flag" in the name.}
}
\value{
Returns \code{dat} with entries in the flag columns converted from
numeric values to text labels (an ordered factor with levels "Pass", "Not
Evaluated", "Suspect", "Fail", "Missing Data")
}
\description{
Will be applied to all columns with "flag" in the name.
}
