% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot_longer.R
\name{qc_pivot_longer}
\alias{qc_pivot_longer}
\title{Pivot flagged sensor string data longer by variable}
\usage{
qc_pivot_longer(dat_wide, qc_tests = NULL)
}
\arguments{
\item{dat_wide}{Data frame of flagged sensor string data in wide format.}

\item{qc_tests}{Quality control tests included in \code{dat_wide}. If
\code{dat_wide} only includes the max flag, use \code{qc_tests = "qc"}.}
}
\value{
Returns \code{dat_wide}, with variables and flags pivoted to a long
format.
}
\description{
Pivot flagged sensor string data longer by variable
}
