% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot_wider.R
\name{qc_pivot_wider}
\alias{qc_pivot_wider}
\title{Pivot flagged sensor string data from long to wide format by variable}
\usage{
qc_pivot_wider(dat_long)
}
\arguments{
\item{dat_long}{Data frame of flagged sensor string data in long format.}
}
\value{
Returns \code{dat_long}, with variables and flags pivoted to a wide
format (a separate column for each variable and variable-qc_test
combination).
}
\description{
Pivot flagged sensor string data from long to wide format by variable
}
