% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{qc_plot_all_tests}
\alias{qc_plot_all_tests}
\title{Title}
\usage{
qc_plot_all_tests(dat, qc_tests = c("climatology", "grossrange"), vars = "all")
}
\arguments{
\item{dat}{placeholder... can be wide or long, but must include columns named
_flag_variable.}

\item{qc_tests}{QC tests to run.}

\item{vars}{Variables to plot}
}
\value{
list of ggplot objects
}
\description{
Title
}
