% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_summarise_flags.R
\name{qc_summarise_flags}
\alias{qc_summarise_flags}
\title{Generate summary table of flags}
\usage{
qc_summarise_flags(dat, qc_tests = NULL, ...)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long or wide format.
Must include at least one column name with the string "_flag_variable".}

\item{qc_tests}{Character vector of quality control tests to apply to
\code{dat}. Defaults to all available tests: \code{qc_tests =
c("climatology", "depth_crosscheck", "grossrange", "rolling_sd", "spike")}.}

\item{...}{Optional argument. Column names (not quoted) from \code{dat} to
use as grouping variables.}
}
\value{
Returns a data frame with the number and percent of observations
assigned each flag value for each variable, qc_test, and groups in
\code{...}.
}
\description{
Generate summary table of flags
}
