% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_all.R
\name{qc_test_all}
\alias{qc_test_all}
\title{Apply multiple quality control tests}
\usage{
qc_test_all(
  dat,
  qc_tests = NULL,
  county = NULL,
  climatology_table = NULL,
  depth_table = NULL,
  grossrange_table = NULL,
  rolling_sd_table = NULL,
  spike_table = NULL,
  join_column = NULL,
  join_column_spike = NULL,
  period_hours = 24,
  max_interval_hours = 2,
  align_window = "center",
  keep_sd_cols = FALSE,
  keep_spike_cols = FALSE
)
}
\arguments{
\item{dat}{Data frame of sensor string data in a wide format.}

\item{qc_tests}{Character vector of quality control tests to apply to
\code{dat}. Defaults to all available tests: \code{qc_tests =
c("climatology", "depth_crosscheck", "grossrange", "rolling_sd", "spike")}.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Used to filter the default \code{climatology_table}. Not
required if there is a \code{county} column in \code{dat}.}

\item{climatology_table}{Data frame with at least 4 columns: \code{variable}:
must match the names of the variables being tested in \code{dat};
\code{month}: numeric values from 1 to 12 for each variable;
\code{season_min}: minimum reasonable value for the corresponding variable
and month; \code{season_max}: maximum reasonable value for the
corresponding variable and month. Optional additional column(s) that is
used to join with \code{dat}. This column must have the same name as the
string \code{join_column}.

Default values are used if \code{climatology_table = NULL}. To see the
default \code{climatology_table}, type \code{subset(thresholds,
  qc_test == "climatology")} in the console.}

\item{depth_table}{Data frame with 1 column: \code{depth_diff_max}.

Default values are used if \code{depth_table = NULL}. To see the
default \code{depth__table}, type \code{subset(thresholds,
  qc_test == "depth_crosscheck")} in the console.}

\item{grossrange_table}{Data frame with at least 7 columns: \code{variable}:
entries must match the names of the variables being tested in \code{dat}.
\code{sensor_type}: type of sensor that recorded the observation;
\code{sensor_min}: minimum value the sensor can record; \code{sensor_max}:
maximum value the sensor can record. \code{user_min}: minimum reasonable
value; \code{user_max}: maximum reasonable value. Optional additional
column(s) that is used to join with \code{dat}. This column must have the
same name as the string \code{join_column}.

Default values are used if \code{grossrange_table = NULL}. To see the
default \code{grossrange_table}, type \code{subset(thresholds,
  qc_test == "grossrange")} in the console.}

\item{rolling_sd_table}{Data frame with at least two columns:
\code{variable}: must match the names of the variables being tested in
\code{dat}. \code{rolling_sd_max}: maximum accepted value for the rolling
standard deviation. Optional additional column that is used to join with
\code{dat}. This column must have the same name as the string
\code{join_column}.

Default values are used if \code{rolling_sd_table = NULL}. To see the
default \code{rolling_sd_table}, type \code{subset(thresholds,
  qc_test == "rolling_sd")} in the console.}

\item{spike_table}{Data frame with at least 3 columns: \code{variable}:
should match the names of the variables being tested in \code{dat}.
\code{spike_low}: maximum acceptable spike value to "Pass", and
\code{spike_high}: maximum acceptable value to be flagged "Suspect/Of
Interest". For variable = sensor_depth_measured_m, the column
\code{sensor_type} is also required.

Optional additional column(s) that is used to join with \code{dat}. This
column must have the same name as the string \code{join_column}.

Default values are used if \code{spike_table = NULL}. To see the
default \code{spike_table}, type \code{subset(thresholds,
  qc_test == "spike")} in the console.}

\item{join_column}{Optional character string of a column name that is in both
\code{dat} and \code{climatology_table}. The specified column will be used
to join the two tables. Default is \code{join_column = NULL}, and the
tables are joined only on the \code{month} and \code{variable} columns.}

\item{join_column_spike}{Optional character string of a column name that is in both
\code{dat} and \code{spike_table}. The specified column will be used to
join the two tables. Default is \code{join_column = NULL}, and the tables
are joined only on the \code{sensor_type} and \code{variable} columns.}

\item{period_hours}{Length of a full cycle in hours. Default assumes a daily
cycle of \code{period_hours = 24}.}

\item{max_interval_hours}{Maximum accepted interval between two observations.
If the interval between two observations is greater than this value, the
rolling standard deviation will be set to \code{NA}. This is important
because for large intervals, the number of observations in
\code{period_hours} will be small. For example, if samples are collected
every 6 hours, only 4 observations would be used to calculate
\code{roll_sd}.}

\item{align_window}{Alignment for the window used to calculate the rolling
standard deviation. Passed to \code{zoo::rollapply()}. Default is
\code{align_window = "center"}. Other options are \code{"right"} (backward
window) and \code{"left"} (forward window).}

\item{keep_sd_cols}{Logical value. If \code{TRUE}, the columns used to
produce the rolling standard deviation (\code{int_sample}, \code{n_sample},
and \code{sd_roll}) are returned in \code{dat}. Default is \code{FALSE}.}

\item{keep_spike_cols}{Logical value. If \code{TRUE}, the columns used to
produce the spike value are returned in \code{dat}. Default is
\code{FALSE}.}
}
\value{
Returns \code{dat} with additional quality control flag columns.
}
\description{
Apply multiple quality control tests
}
