% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_depth_crosscheck.R
\name{qc_test_depth_crosscheck}
\alias{qc_test_depth_crosscheck}
\title{Apply the depth crosscheck test}
\usage{
qc_test_depth_crosscheck(
  dat,
  depth_table = NULL,
  county = NULL,
  keep_depth_cols = FALSE
)
}
\arguments{
\item{dat}{Data frame of sensor string data in a wide format.}

\item{depth_table}{Data frame with 1 column: \code{depth_diff_max}.

Default values are used if \code{depth_table = NULL}. To see the
default \code{depth__table}, type \code{subset(thresholds,
  qc_test == "depth_crosscheck")} in the console.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Used to filter the default \code{depth_crosscheck_table}.
Not required if there is a \code{county} column in \code{dat} or if
\code{depth_table} is provided.}

\item{keep_depth_cols}{Logical value. If \code{TRUE}, the columns used to
evaluate the difference between measured depth and estimated depth
(\code{min_measured} and \code{abs_diff}) are returned in \code{dat}.
Default is \code{FALSE}.}
}
\value{
Returns \code{dat} in a wide format, with depth crosscheck flag
column named "depth_crosscheck_flag_value".
}
\description{
Sensor depth at low tide is estimated from tide charts and the position of
each sensor on the string. The estimated depth is recorded in in the
deployment log and included in the compiled data in the column
\code{sensor_depth_at_low_tide_m}.
}
\details{
Some sensors are capable of recording depth (pressure). When this is the
case, the minimum sensor depth recorded (i.e., low tide) is compared to the
estimated sensor depth at low tide. In the case of large discrepancies, the
depth_crosscheck_flag will indicate that the observation is "Suspect / Of
Interest". In many cases, the measured sensor depth should be considered more
accurate than the estimated depth; however, the measured depth should be
carefully evaluated.

Note that if there are multiple sensors on the string, \emph{only} the sensor(s)
with measured depth values will be evaluated with the depth crosscheck test.
If the depth crosscheck test identifies the estimated depth as "Suspect", the
other estimated sensor depths should also be considered suspect. For example,
if the string was moored in an area 10 m deeper than anticipated, all sensors
will likely be 10 m deeper than recorded in the
\code{sensor_depth_at_low_tide_m} column.
}
