% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_flat_line.R
\name{qc_test_flat_line}
\alias{qc_test_flat_line}
\title{Apply the flat line test}
\usage{
qc_test_flat_line(dat, flat_line_table = NULL, keep_lag_cols = FALSE)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{flat_line_table}{text}

\item{keep_lag_cols}{text}
}
\value{
Format depends on keep_lag_cols
}
\description{
Only checks if values are \emph{equal} (no tolerance).
}
\details{
Flag of 2 assigned until can evaluate if it fails.
}
\seealso{
Other tests: 
\code{\link{qc_test_climatology}()},
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_rolling_sd}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
