% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_gap.R
\name{qc_test_gap}
\alias{qc_test_gap}
\title{Add flag column for the gap test}
\usage{
qc_test_gap(dat, variable, threshold)
}
\arguments{
\item{dat}{Data.frame with at least one column \code{timestamp}.}

\item{variable}{Variable to apply the test to.}

\item{threshold}{Gap that fails test.}
}
\value{
placeholder for now
}
\description{
Add flag column for the gap test
}
