% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_location.R
\name{qc_test_location}
\alias{qc_test_location}
\title{Add flag column for the location test}
\usage{
qc_test_location(
  dat,
  latmax = 47.146147,
  latmin = 43.336527,
  longmax = -59.496715,
  longmin = -66.457067
)
}
\arguments{
\item{dat}{Data.frame with at least one column \code{timestamp}.}

\item{latmax}{maximum latitude}

\item{latmin}{minimum latitude}

\item{longmax}{maximum longitude}

\item{longmin}{minimum longitude}
}
\value{
placeholder for now
}
\description{
Add flag column for the location test
}
