% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_rate_of_change.R
\name{qc_test_rate_of_change}
\alias{qc_test_rate_of_change}
\title{Apply the gross range test}
\usage{
qc_test_rate_of_change(
  dat,
  period_hours = 24,
  rate_of_change_table = NULL,
  align_window = "center",
  keep_sd_cols = FALSE
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{period_hours}{Length of a full cycle in hours. Default assumes a daily
cycle of \code{period_hours = 24}.}

\item{rate_of_change_table}{Data frame with two columns: \code{variable}:
must match the names of the variables being tested in \code{dat}.
\code{stdev_max}: minimum accepted value for the rolling standard
deviation. Default values are used if \code{rate_of_change_table = NULL}.}

\item{align_window}{Alignment for the window used to calculate the rolling
standard deviation. Passed to \code{zoo:rollapply()}. Default is
\code{align_window = "center"}. Other options are \code{"right"} (backward
window) and \code{"left"} (forward window).}

\item{keep_sd_cols}{Logical value. If \code{TRUE}, the columns used to
produce the rolling standard deviation (int_sample, n_sample, and sd_roll)
are returned in \code{dat}. Default is \code{FALSE}.}
}
\value{
Returns \code{dat} in a wide format, with rate of change flag columns
for each variable in the form "rate_of_change_flag_variable".
}
\description{
Apply the gross range test
}
\seealso{
Other tests: 
\code{\link{qc_test_climatology}()},
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
