% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_v2r_for_hil.R
\name{qc_vr2_for_hil}
\alias{qc_vr2_for_hil}
\title{Identify redundant temperature data from VR2 sensors}
\usage{
qc_vr2_for_hil(dat)
}
\arguments{
\item{dat}{Data frame of compiled sensor string data for a single deployment.}
}
\value{
A data frame row that can be appended to
water_quality_human_in_loop.
}
\description{
VR2AR temperature sensors are substantially less accurate than other
temperature sensors. Starting in 2025, VR2AR units will be used only for
acoustic detections and the release. A hobo sensor will be attached to each
VR2AR to record bottom temperature. The corresponding VR2AR temperature will
be assigned a human in the loop flag of 4 for CMAR records, and will \emph{not} be
published on the Open Data Portal
}
