% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tables.R
\docType{data}
\name{threshold_tables}
\alias{threshold_tables}
\title{Example of compiled Hobo data in wide format, as exported from
compile_HOBO_data()}
\format{
Default thresholds for adding flags to water quality data. A list
with a separate data frame for each test: climatology_table and
grossrange_table.
}
\source{
Add references here
}
\usage{
threshold_tables
}
\description{
Example of compiled Hobo data in wide format, as exported from
compile_HOBO_data()
}
\keyword{datasets}
