% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_tables.R
\docType{data}
\name{thresholds}
\alias{thresholds}
\title{Threshold tables}
\format{
Default thresholds for adding flags to water quality data.
}
\source{
Add references here
}
\usage{
thresholds
}
\description{
Threshold tables
}
\keyword{datasets}
