% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{ggplot_flags}
\alias{ggplot_flags}
\title{Create ggplot for one qc_test and variable}
\usage{
ggplot_flags(dat, qc_test, var, ncol = NULL)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long format.
Must include a column named with the string "_flag_value".}

\item{qc_test}{qc test to plot.}

\item{var}{variable to plot.}

\item{ncol}{Number of columns for faceted plots.}
}
\value{
ggplot object
}
\description{
Create ggplot for one qc_test and variable
}
