% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{ggplot_flags}
\alias{ggplot_flags}
\title{Create ggplot for one qc_test and variable}
\usage{
ggplot_flags(dat, qc_test, var, ncol = NULL, flag_title = TRUE)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long format. Must
include a column named with the string "_flag_value".}

\item{qc_test}{qc test to plot.}

\item{var}{variable to plot.}

\item{ncol}{Number of columns for faceted plots.}

\item{flag_title}{Logical argument indicating whether to include a ggtitle of
the qc test and variable plotted.}
}
\value{
Returns a ggplot object; a figure for \code{qc_test} and \code{var}.
Points are coloured by the flag value and panels are faceted by depth and
sensor.
}
\description{
Create ggplot for one qc_test and variable
}
