% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot_wider.R
\name{pivot_flags_wider}
\alias{pivot_flags_wider}
\title{Complete pivot_wider of flagged sensor string data}
\usage{
pivot_flags_wider(dat_long, var)
}
\arguments{
\item{dat_long}{Data frame of flagged sensor string data with the variables
in a long format and flags in wide format.}

\item{var}{Variable to pivot.}
}
\value{
Returns \code{dat_long} with the variable and qc_test flag columns
pivoted to a wide format.
}
\description{
Complete pivot_wider of flagged sensor string data
}
