% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_all.R
\name{qc_assign_max_flag}
\alias{qc_assign_max_flag}
\title{Assign a single QC flag value for each observation}
\usage{
qc_assign_max_flag(dat)
}
\arguments{
\item{dat}{wide data frame with all flag columns}
}
\value{
dat with only 1 flag column
}
\description{
Assign a single QC flag value for each observation
}
