% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calculate_thresholds.R
\name{qc_calculate_rolling_sd_thresholds}
\alias{qc_calculate_rolling_sd_thresholds}
\title{Calculate rolling standard deviation thresholds}
\usage{
qc_calculate_rolling_sd_thresholds(dat, var, prob = 0.95)
}
\arguments{
\item{dat}{data.frame. Must include column sd_roll.}

\item{var}{variable to calculate thresholds for.}

\item{prob}{sent to quantile function}
}
\value{
tibble with
}
\description{
The rolling standard deviation thresholds are calculated from historical data
as the upper quartile of historical observations.
}
\details{
This function does not automatically group by any variables. To calculate the
thresholds by group (e.g., county) use
\code{dplyr::group_by(group_variable)}, and send the results to
\code{qc_calculate_rolling_sd_thresholds()}.
}
