% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_calculate_thresholds.R
\name{qc_calculate_rolling_sd_thresholds}
\alias{qc_calculate_rolling_sd_thresholds}
\title{Calculate rolling standard deviation thresholds}
\usage{
qc_calculate_rolling_sd_thresholds(
  dat,
  var,
  stat = NULL,
  prob = 0.95,
  n_sd = 3
)
}
\arguments{
\item{dat}{data frame. Must include column sd_roll.}

\item{var}{variable to calculate thresholds for.}

\item{stat}{Statistic to calculate for the threshold. Options are: 1.
"quartile", which uses the quantile function to calculate the value of the
\code{prob} quartile. Used for variables with a right-tailed distribution
for rolling standard deviation.
\enumerate{
\item "mean_sd", which calculates the threshold as the mean + \code{n_sd}
standard deviations. Used for variables with normal distributions for
rolling standard deviation.
}}

\item{prob}{sent to quantile function. Only required when \code{stat =
"quartile"}.}

\item{n_sd}{Number of standard deviations to use. Default is \code{n_sd = 3}.
Only required when \code{stat = "mean_sd"}.}
}
\value{
tibble with
}
\description{
The rolling standard deviation thresholds are calculated from historical data
as the upper quartile of historical observations.
}
\details{
This function does not automatically group by any variables. To calculate the
thresholds by group (e.g., county) use
\code{dplyr::group_by(group_variable)}, and send the results to
\code{qc_calculate_rolling_sd_thresholds()}.
}
