% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pivot.R
\name{qc_pivot_longer}
\alias{qc_pivot_longer}
\title{Title}
\usage{
qc_pivot_longer(dat_wide, qc_tests = c("climatology", "grossrange"))
}
\arguments{
\item{dat_wide}{placeholder}

\item{qc_tests}{QC tests to run.}
}
\value{
placeholder
}
\description{
Title
}
