% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{qc_plot_flags}
\alias{qc_plot_flags}
\title{Plot sensor data coloured by flag value}
\usage{
qc_plot_flags(
  dat,
  qc_tests = c("climatology", "grossrange"),
  vars = "all",
  labels = TRUE
)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long or wide format.
Must include at least one column name with the string "_flag_variable".}

\item{qc_tests}{QC tests to run.}

\item{vars}{Character vector of variables to plot.}

\item{labels}{Logical argument indicating whether to convert numeric flag
values to text labels for the legend.}
}
\value{
Returns a list of ggplot objects; one figure for each test in
\code{qc_tests} and variable in \code{vars}, faceted by depth and sensor.
}
\description{
Plot sensor data coloured by flag value
}
