% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_flag_vars.R
\name{qc_plot_flags}
\alias{qc_plot_flags}
\title{Plot sensor data coloured by flag value}
\usage{
qc_plot_flags(
  dat,
  qc_tests = c("climatology", "depth_crosscheck", "grossrange", "rolling_sd", "spike"),
  vars = "all",
  labels = TRUE,
  ncol = NULL,
  flag_title = TRUE
)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in long or wide format.
Must include at least one column name with the string "_flag_variable".}

\item{qc_tests}{Character vector of quality control tests to apply to
\code{dat}. Defaults to all available tests: \code{qc_tests =
  c("climatology", "depth_crosscheck", "grossrange", "rolling_sd", "spike")}}

\item{vars}{Character vector of variables to plot. Default is \code{vars =
"all"}, which will make a plot for each recognized variable in \code{dat}.}

\item{labels}{Logical argument indicating whether to convert numeric flag
values to text labels for the legend.}

\item{ncol}{Number of columns for faceted plots.}

\item{flag_title}{Logical argument indicating whether to include a ggtitle of
the qc test and variable plotted.}
}
\value{
Returns a list of ggplot objects; one figure for each test in
\code{qc_tests} and variable in \code{vars}. Points are coloured by the
flag value and panels are faceted by depth and sensor. faceted by depth and
sensor.
}
\description{
Plot sensor data coloured by flag value
}
