% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_all.R
\name{qc_test_all}
\alias{qc_test_all}
\title{Run specified quality control tests}
\usage{
qc_test_all(
  dat,
  qc_tests = c("climatology", "grossrange"),
  climatology_table = NULL,
  seasons_table = NULL,
  grossrange_table = NULL
)
}
\arguments{
\item{dat}{Data in a wide format. Add more here}

\item{qc_tests}{QC tests to run.}

\item{climatology_table}{Data frame with 4 columns: \code{variable}: should
match the names of the variables being tested in \code{dat}. \code{season}:
there should be an entry of "winter", "spring", "summer", and "fall" for
each variable. \code{season_min}: minimum reasonable value for the
corresponding variable during the corresponding season. \code{season_max}:
maximum reasonable value for the corresponding variable during the
corresponding season.

Default values are used if \code{climatology_table = NULL}. To see the
default \code{climatology_table}, type
\code{threshold_tables$climatology_table} in the console.}

\item{seasons_table}{Data frame with 2 columns: \code{month}: numeric value
of the month and \code{season}, the corresponding season (entries of
"winter", "spring", "summer", and "fall"). Default table is used if
\code{seasons = NULL}. To see the default values, type
\code{threshold_tables$seasons} in the console.}

\item{grossrange_table}{Data frame with 6 columns: \code{variable}: should
match the names of the variables being tested in \code{dat}.
\code{sensor_make}: this may change \code{sensor_min}: minimum value the
sensor can record. \code{sensor_max}: maximum value the sensor can record.
\code{user_min}: minimum reasonable value; smaller values are "of
interest". \code{user_max}: maximum reasonable value; larger values are "of
interest".

Default is \code{grossrange_table = NULL}, which uses default values. To
see the default \code{grossrange_table}, type
\code{threshold_tables$grossrange_table}.}
}
\value{
Tibble of \code{dat} with quality control flag columns.
}
\description{
Run specified quality control tests
}
