% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_all.R
\name{qc_test_all}
\alias{qc_test_all}
\title{Apply multiple quality control tests}
\usage{
qc_test_all(
  dat,
  qc_tests = c("climatology", "grossrange", "spike"),
  climatology_table = NULL,
  grossrange_table = NULL,
  spike_table = NULL,
  county,
  message = TRUE
)
}
\arguments{
\item{dat}{Data frame of sensor string data in a wide format.}

\item{qc_tests}{Character vector of quality control tests to apply to
\code{dat}. Defaults to all available tests: \code{qc_tests =
  c("climatology", "grossrange", "spike")}}

\item{climatology_table}{Data frame with 4 columns: \code{variable}: must
match the names of the variables being tested in \code{dat}; \code{month}:
numeric values from 1 to 12 for each variable; \code{season_min}: minimum
reasonable value for the corresponding variable and month;
\code{season_max}: maximum reasonable value for the corresponding variable
and month.

Default values are used if \code{climatology_table = NULL}. To see the
default \code{climatology_table}, type \code{subset(threshold_tables,
  qc_test == "climatology")} in the console.}

\item{grossrange_table}{Data frame with 8 columns: \code{variable}: must
match the names of the variables being tested in \code{dat}.
\code{sensor_type}: type of sensor that recorded the observation;
\code{sensor_min}: minimum value the sensor can record; \code{sensor_max}:
maximum value the sensor can record. \code{user_min}: minimum reasonable
value; \code{user_max}: maximum reasonable value.

Default values are used if \code{grossrange_table = NULL}. To see the
default \code{grossrange_table}, type \code{subset(threshold_tables,
  qc_test == "grossrange")} in the console.}

\item{spike_table}{Data frame with 3 columns: \code{variable}: should match
the names of the variables being tested in \code{dat}.

Default is \code{spike_table = NULL}, which uses default values. To see the
default \code{spike_table}, type \code{threshold_tables$spike_table}.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Required if the default \code{climatology_table} is used.}

\item{message}{Logical argument indicating whether to display a message if
any \code{user_min} < \code{sensor_min} and/or \code{user_max} >
\code{sensor_max} (i.e., flag of 3 cannot be applied).}
}
\value{
Returns \code{dat} with additional quality control flag columns.
}
\description{
Apply multiple quality control tests
}
