% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_climatology.R
\name{qc_test_climatology}
\alias{qc_test_climatology}
\title{Add flag column for the climatology test}
\usage{
qc_test_climatology(dat)
}
\arguments{
\item{dat}{Data.frame with at least one column \code{timestamp}.}
}
\value{
placeholder for now
}
\description{
Add flag column for the climatology test
}
