% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_climatology.R
\name{qc_test_climatology}
\alias{qc_test_climatology}
\title{Apply the climatology test}
\usage{
qc_test_climatology(dat, climatology_table = NULL, county = NULL)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{climatology_table}{Data frame with 4 columns: \code{variable}: must
match the names of the variables being tested in \code{dat}; \code{month}:
numeric values from 1 to 12 for each variable; \code{season_min}: minimum
reasonable value for the corresponding variable and month;
\code{season_max}: maximum reasonable value for the corresponding variable
and month.

Default values are used if \code{climatology_table = NULL}. To see the
default \code{climatology_table}, type \code{subset(threshold_tables,
  qc_test == "climatology")} in the console.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Required if the default \code{climatology_table} is used.}
}
\value{
Returns \code{dat} in a wide format, with climatology flag columns
for each variable in the form "climatology_flag_variable".
}
\description{
Apply the climatology test
}
\seealso{
Other tests: 
\code{\link{qc_test_flat_line}()},
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_rolling_sd}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
