% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_climatology.R
\name{qc_test_climatology}
\alias{qc_test_climatology}
\title{Add flag column for the climatology test}
\usage{
qc_test_climatology(dat, climatology_table = NULL, seasons_table = NULL)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{climatology_table}{Data frame with 4 columns: \code{variable}: should
match the names of the variables being tested in \code{dat}. \code{season}:
there should be an entry of "winter", "spring", "summer", and "fall" for
each variable. \code{season_min}: minimum reasonable value for the
corresponding variable during the corresponding season. \code{season_max}:
maximum reasonable value for the corresponding variable during the
corresponding season.

Default values are used if \code{climatology_table = NULL}. To see the
default \code{climatology_table}, type
\code{threshold_tables$climatology_table} in the console.}

\item{seasons_table}{Data frame with 2 columns: \code{month}: numeric value
of the month and \code{season}, the corresponding season (entries of
"winter", "spring", "summer", and "fall"). Default table is used if
\code{seasons = NULL}. To see the default values, type
\code{threshold_tables$seasons} in the console.}
}
\value{
placeholder for now
}
\description{
Add flag column for the climatology test
}
\seealso{
Other tests: 
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
