% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_depth_crosscheck.R
\name{qc_test_depth_crosscheck}
\alias{qc_test_depth_crosscheck}
\title{Test the recorded sensor depth matches measured sensor depth.}
\usage{
qc_test_depth_crosscheck(
  dat,
  depth_table = NULL,
  county = NULL,
  keep_depth_cols = FALSE
)
}
\arguments{
\item{dat}{Dat}

\item{depth_table}{TBD}

\item{county}{TBD}

\item{keep_depth_cols}{TBD}
}
\value{
TBD
}
\description{
Test the recorded sensor depth matches measured sensor depth.
}
