% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_grossrange.R
\name{qc_test_grossrange}
\alias{qc_test_grossrange}
\title{Add flag column for the grossrange test}
\usage{
qc_test_grossrange(dat)
}
\arguments{
\item{dat}{Data.frame with at least one column \code{timestamp}.}
}
\value{
placeholder for now
}
\description{
Add flag column for the grossrange test
}
