% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_grossrange.R
\name{qc_test_grossrange}
\alias{qc_test_grossrange}
\title{Apply the gross range test}
\usage{
qc_test_grossrange(dat, grossrange_table = NULL, county = NULL, message = TRUE)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{grossrange_table}{Data frame with 8 columns: \code{variable}: must
match the names of the variables being tested in \code{dat}.
\code{sensor_type}: type of sensor that recorded the observation;
\code{sensor_min}: minimum value the sensor can record; \code{sensor_max}:
maximum value the sensor can record. \code{user_min}: minimum reasonable
value; \code{user_max}: maximum reasonable value.

Default values are used if \code{grossrange_table = NULL}. To see the
default \code{grossrange_table}, type \code{subset(threshold_tables,
  qc_test == "grossrange")} in the console.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Required if the default \code{grossrange_table} is used.}

\item{message}{Logical argument indicating whether to display a message if
any \code{user_min} < \code{sensor_min} and/or \code{user_max} >
\code{sensor_max} (i.e., flag of 3 cannot be applied).}
}
\value{
Returns \code{dat} in a wide format, with grossrange flag columns for
each variable in the form "grossrange_flag_variable".
}
\description{
The variable value must be greater than the threshold to trigger flag (e.g.,
value > sensor_max is assigned a flag of 4; value == sensor_max is assigned a
flag of 3).
}
\details{
The sensor_type is reassigned as "hobo do" if sensor_type = hobo and
do_concentration is not NA so that the data can be joined with the
\code{grossrange_table} by \code{sensor_type} and \code{variable}.
}
\seealso{
Other tests: 
\code{\link{qc_test_climatology}()},
\code{\link{qc_test_flat_line}()},
\code{\link{qc_test_rolling_sd}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
