% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_human_in_loop.R
\name{qc_test_human_in_loop}
\alias{qc_test_human_in_loop}
\title{Apply human in the loop flags and comments}
\usage{
qc_test_human_in_loop(
  dat,
  human_in_loop_table = NULL,
  qc_tests = c("climatology", "grossrange", "rolling_sd", "spike")
)
}
\arguments{
\item{dat}{Data frame of flagged sensor string data in wide format.}

\item{human_in_loop_table}{Data table with information required to assign the
human in the loop flags. Must include the following columns: station,
depl_range, variable, sensor_serial number (if \code{NA}, all sensors that
recorded variable will be flagged), timestamp_prompt (indicates whether
flags should be applied "between" the timestamp_utc_min
and    timestamp_utc_max or "outside" these timestamps),
timestamp_utc_min,    timestamp_utc_max, qc_test_column (qc flag column that
is being upgraded), qc_flag_value (existing flag value for qc_test_column),
human_in_loop_flag_value, and human_in_loop_comment (will be included in
the dataset).}

\item{qc_tests}{Character vector of quality control tests that have been
applied to \code{dat}. Passed to \code{qc_pivot_longer()}. Default is:
\code{qc_tests = c("climatology", "grossrange", "rolling_sd", "spike")}.}
}
\value{
Returns \code{dat} with a \code{human_in_loop_flag} column for each
variable and a corresponding \code{hil_comment} column.
}
\description{
Apply human in the loop flags and comments
}
