% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_rolling_sd.R
\name{qc_test_rolling_sd}
\alias{qc_test_rolling_sd}
\title{Apply the rolling standard deviation test}
\usage{
qc_test_rolling_sd(
  dat,
  period_hours = 24,
  rolling_sd_table = NULL,
  align_window = "center",
  keep_sd_cols = FALSE,
  log_sd = FALSE,
  offset = 0.005
)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{period_hours}{Length of a full cycle in hours. Default assumes a daily
cycle of \code{period_hours = 24}.}

\item{rolling_sd_table}{Data frame with two columns: \code{variable}: must
match the names of the variables being tested in \code{dat}.
\code{stdev_max}: minimum accepted value for the rolling standard
deviation. Default values are used if \code{rolling_sd_table = NULL}.}

\item{align_window}{Alignment for the window used to calculate the rolling
standard deviation. Passed to \code{zoo:rollapply()}. Default is
\code{align_window = "center"}. Other options are \code{"right"} (backward
window) and \code{"left"} (forward window).}

\item{keep_sd_cols}{Logical value. If \code{TRUE}, the columns used to
produce the rolling standard deviation (int_sample, n_sample, and sd_roll)
are returned in \code{dat}. Default is \code{FALSE}.}

\item{log_sd}{Logical value. If \code{TRUE}, the natural log of the standard
deviation is used. (Because more normal distribution. Make sure to use
appropriate threshold.)}

\item{offset}{default 0.005}
}
\value{
Returns \code{dat} in a wide format, with rolling standard deviation
flag columns for each variable in the form "rolling_sd_flag_variable".
}
\description{
\code{sd_roll} is rounded to 2 decimal places.
}
\seealso{
Other tests: 
\code{\link{qc_test_climatology}()},
\code{\link{qc_test_flat_line}()},
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_spike}()}
}
\concept{tests}
