% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_test_spike.R
\name{qc_test_spike}
\alias{qc_test_spike}
\title{Add flag columns for the spike test}
\usage{
qc_test_spike(dat, spike_table = NULL)
}
\arguments{
\item{dat}{Data frame of sensor string data in wide format.}

\item{spike_table}{Data frame with 3 columns: \code{variable}: should match
the names of the variables being tested in \code{dat}.

Default is \code{spike_table = NULL}, which uses default values. To see the
default \code{spike_table}, type \code{threshold_tables$spike_table}.}
}
\value{
placeholder for now
}
\description{
Might want to make this dependent on sample rate (sensor type)
}
\seealso{
Other tests: 
\code{\link{qc_test_climatology}()},
\code{\link{qc_test_grossrange}()},
\code{\link{qc_test_rate_of_change}()}
}
\concept{tests}
